generate suurls for GET conv when xid present
This was added so we could prototype an integration.
The long-term solution may be to use OAuth2.

Problems:
This stores appends suzinvites to the DB each time it is called, so
if we're still using this down the line, we should switch to using
HMAC-based urls to link xid identities.
The single-use property of these isn't actually needed for this
use-case.