adjust mongo queries.
MongoLab says the queries weren't using zid index. This seems likely to
be true since it was taking an average of 177ms, and it takes an
average of 220 ms to return the whole table (every zid).

This is an attempt to make it super obvious to mongo that it does in
fact have the index it needs to make this fast.